+++ tests/test_models.py
@@ -57,8 +57,65 @@ class TestProperty:
 
 class TestPropertyValue:
     def test_create__status(self):
-        p = PropertyValue.create("status", value="Done")
-        p_dict = p.get()
-        assert p.id is None
-        assert p.type == "status"
+        pv = PropertyValue.create("status", value="Done")
+        p_dict = pv.get()
+        assert pv.id is None
+        assert pv.type == "status"
         assert p_dict["status"]["name"] == "Done"
+
+    def test_create__relation(self):
+        pv = PropertyValue.create("relation", value=[LinkTo.create(page_id="04262843082a478d97f741948a32613c")])
+        p_dict = pv.get()
+        assert pv.id is None
+        assert pv.type == "relation"
+        assert pv.has_more is False
+        assert p_dict["relation"][0]["id"] == "04262843082a478d97f741948a32613c"
+
+
+class TestBlock:
+    def test_get__heading_1(self, no):
+        block_id = "15a5790980db4e8798b9b7801385afbb"
+        block = no.blocks.get(block_id)
+        assert isinstance(block.obj, Block)
+        assert isinstance(block.obj.text, RichTextArray)
+        assert block.obj.type == "heading_1"
+        assert block.obj.simple == "Block with heading 1 type"
+        assert str(block.obj) == "# Block with heading 1 type"
+        assert block.obj.has_children is False
+        assert block.obj.is_toggleable is False
+
+    def test_get__t_heading_1(self, no):
+        block_id = "a62985febcac499f95e5b59643df6180"
+        block = no.blocks.get(block_id)
+        assert isinstance(block.obj, Block)
+        assert isinstance(block.obj.text, RichTextArray)
+        assert block.obj.type == "heading_1"
+        assert block.obj.simple == "Block with Toggle Heading 1 type with no children"
+        assert str(block.obj) == "# Block with Toggle Heading 1 type with no children"
+        assert block.obj.has_children is False
+        assert block.obj.is_toggleable is True
+
+    def test_get__t_heading_2(self, no):
+        block_id = "4b265c74e7644affad13a3820e208b78"
+        block = no.blocks.get(block_id)
+        assert isinstance(block.obj, Block)
+        assert isinstance(block.obj.text, RichTextArray)
+        assert block.obj.type == "heading_2"
+        assert block.obj.simple == "Block with Toggle Heading 2 type with children"
+        assert str(block.obj) == "## Block with Toggle Heading 2 type with children"
+        assert block.obj.has_children is True
+        assert block.obj.is_toggleable is True
+
+    def test_create__heading_1(self):
+        b = Block.create("hello there", type_="heading_1")
+        b_dict = b.get()
+        assert b.id == ""
+        assert b.type == "heading_1"
+        assert b_dict["heading_1"]["is_toggleable"] is False
+
+    def test_create__t_heading_2(self):
+        b = Block.create("hello there", type_="heading_2", is_toggleable=True)
+        b_dict = b.get()
+        assert b.id == ""
+        assert b.type == "heading_2"
+        assert b_dict["heading_2"]["is_toggleable"] is True